requirements.txt
torch>=2.7.0,<2.8.0
# TODO: xlm-core latest, update to pip
#git+https://github.com/dhruvdcoder/xlm-core.git@f7a20eae5965a04513079245d60959098c9f0f0a
-e /scratch3/workspace/dhruveshpate_umass_edu-idlm/xlm-core
hydra-core
lightning==2.5.2
fsspec>=2025.3.0 # https://github.com/Lightning-AI/pytorch-lightning/pull/20780
# --index-url https://download.pytorch.org/whl/cu124 ; torch==2.5.1+cu124
# you can change the cuda version but make sure to use torch 2.5.1
jupyter
hydra-colorlog
hydra-joblib-launcher
hydra-submitit-launcher
datasets >= 3.3.2, < 4.0.0 # was using 2.15.0 earlier due to some bug. Then used 3.3.2 successfully. With 4.0.0, one cannot use script based datasets but lm1b is script based as of Sep 5, 2025.
transformers
more-itertools
torchdata>=0.9.0
rich
python-dotenv
jaxtyping
tensorboard
torch-ema
pandas
# logging
wandb
tensorboard
# visualization
matplotlib
seaborn
networkx
# molecule generation
pot>=0.9.5
# scikit-learn==1.2.2 install this after due to soft bionemo-moco conflict
einops
safe-mol==0.1.14
pytdc==0.4.1
easydict
openbabel-wheel==3.1.1.22